refactor: rename ThriftUtils to SubTaskModelDecoder and move to task package
The class is a pure protobuf sub-task decoder with no Thrift content; the
name was a leftover from the pre-gRPC era. Relocate it next to its only
consumers in the task package and update both call sites.
